Rubick/test

See merge request orderlynetwork/orderly-v2/contract-poc-demo!29

+++ test/mock/VaultCrossChainManagerMock.sol
@@ -5,20 +5,10 @@ import "../../src/interface/IVaultCrossChainManager.sol";
 import "openzeppelin-contracts/contracts/access/Ownable.sol";
 import "crosschain/interface/IOrderlyCrossChain.sol";
 
-contract VaultCrossChainManagerMock is
-    IOrderlyCrossChainReceiver,
-    IVaultCrossChainManager,
-    Ownable
-{
-    function receiveMessage(
-        bytes memory payload,
-        uint srcChainId,
-        uint dstChainId
-    ) external override {}
+contract VaultCrossChainManagerMock is IOrderlyCrossChainReceiver, IVaultCrossChainManager, Ownable {
+    function receiveMessage(bytes memory payload, uint256 srcChainId, uint256 dstChainId) external override {}
 
-    function withdraw(
-        OrderlyCrossChainMessage.MessageV1 memory message
-    ) external override {}
+    function withdraw(OrderlyCrossChainMessage.MessageV1 memory message) external override {}
 
     function deposit(VaultTypes.VaultDeposit memory data) external override {}
